LogDataクラスの移動
